safari-600.7-branch/LayoutTests/platform/mac/accessibility/element-focus-expected.txt

link
text
foo
This tests that any element that supports focus has a writable AXFocused

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS body.childAtIndex(0).childAtIndex(0).role is 'AXRole: AXLink'
PASS body.childAtIndex(0).childAtIndex(0).isAttributeSettable('AXFocused') is true
PASS body.childAtIndex(1).role is 'AXRole: AXStaticText'
PASS body.childAtIndex(1).isAttributeSettable('AXFocused') is true
PASS body.childAtIndex(2).role is 'AXRole: AXGroup'
PASS body.childAtIndex(2).isAttributeSettable('AXFocused') is true
PASS successfullyParsed is true

TEST COMPLETE
